Made the buffer position of the PostScript printer 32-bit as it should have been.

+++ src/printer/prt_ps.c
@@ -86,7 +86,7 @@ typedef struct
     wchar_t	filename[260];
 
     char	buffer[POSTSCRIPT_BUFFER_LENGTH];
-    uint16_t	buffer_pos;
+    uint32_t	buffer_pos;
 } ps_t;
 
 static void
